moved some test utils to LndTest
