+++ src/LndClient/LndTest.hs
@@ -19,17 +19,40 @@ module LndClient.LndTest
     spawnLinkChannelWatcher,
     spawnLinkInvoiceWatcher,
     watchDefaults,
+    receiveClosedChannels,
+    receiveActiveChannel,
+    receiveInvoice,
+    closeAllChannels,
+    setupOneChannel,
   )
 where
 
+import LndClient.Data.AddInvoice as AddInvoice
+  ( AddInvoiceRequest (..),
+    AddInvoiceResponse (..),
+  )
+import LndClient.Data.Channel as Channel (Channel (..))
+import LndClient.Data.ChannelPoint as ChannelPoint (ChannelPoint (..))
+import LndClient.Data.CloseChannel
+  ( ChannelCloseSummary (..),
+    CloseChannelRequest (..),
+  )
+import LndClient.Data.GetInfo (GetInfoResponse (..))
 import qualified LndClient.Data.GetInfo as LND (GetInfoResponse (..))
-import LndClient.Data.Invoice as LND (Invoice (..))
+import LndClient.Data.Invoice as Invoice (Invoice (..))
+import LndClient.Data.ListChannels as LC (ListChannelsRequest (..))
 import qualified LndClient.Data.NewAddress as LND
   ( NewAddressResponse (..),
+  )
+import LndClient.Data.OpenChannel as OpenChannel
+  ( OpenChannelRequest (..),
   )
 import LndClient.Data.Peer
   ( ConnectPeerRequest (..),
     LightningAddress (..),
+  )
+import LndClient.Data.SendPayment as SendPayment
+  ( SendPaymentRequest (..),
   )
 import LndClient.Data.SubscribeChannelEvents
 import LndClient.Data.SubscribeInvoices
@@ -37,6 +60,7 @@ import LndClient.Data.SubscribeInvoices
   )
 import LndClient.Import
 import qualified LndClient.RPC.Katip as LND
+import LndClient.Util as Util
 import LndClient.Watcher as Watcher
 import qualified LndGrpc as GRPC
 import qualified Network.Bitcoin as BTC (Client, getClient)
@@ -212,6 +236,162 @@ syncWallets = this 0
           if LND.syncedToChain x0 && LND.syncedToChain x1
             then pure $ Right ()
             else liftIO (delay 1000000) >> this (attempt + 1)
+
+receiveClosedChannels ::
+  (LndTest m) =>
+  [ChannelPoint] ->
+  TChan ((), ChannelEventUpdate) ->
+  m (Either LndError ())
+receiveClosedChannels = this 0
+  where
+    this _ [] _ =
+      pure $ Right ()
+    this 30 _ _ =
+      pure
+        $ Left
+        $ LndError "receiveClosedChannels - exceeded"
+    this (attempt :: Integer) cps0 cq = do
+      x0 <- readTChanTimeout (MicroSecondsDelay 1000000) cq
+      let x = channelEvent . snd <$> x0
+      $(logTM) InfoS $ logStr $
+        "receiveClosedChannels - got " <> (show x :: Text)
+      case x of
+        Just (ChannelEventUpdateChannelClosedChannel res) ->
+          case filter (/= chPoint res) cps0 of
+            [] -> pure $ Right ()
+            cps -> mine1 >> this (attempt + 1) cps cq
+        _ ->
+          mine1 >> this (attempt + 1) cps0 cq
+
+closeAllChannels :: (LndTest m) => m ()
+closeAllChannels = do
+  cq <- dupChannelTChan Alice
+  mq <- dupChannelTChan Bob
+  $(logTM) InfoS "CloseAllChannels - closing channels ..."
+  alice <- getLndEnv Alice
+  cs <-
+    liftLndResult
+      =<< LND.listChannels
+        alice
+        (ListChannelsRequest True False False False Nothing)
+  let cps = Channel.channelPoint <$> cs
+  mapM_
+    ( \cp ->
+        Util.spawnLink $
+          LND.closeChannel
+            --
+            -- TODO : investigate why it throws empty gRPC
+            -- response error and as consequence
+            -- probably it's how subscription terminates
+            -- when channel is completely closed
+            --
+            -- but bad thing is that callback sometimes
+            -- is never called, that's why it's not used there
+            -- we are getting TChan events instead from
+            -- previous opened subscription
+            --
+            (const $ return ())
+            alice
+            (CloseChannelRequest cp False Nothing Nothing Nothing)
+    )
+    cps
+  liftLndResult =<< receiveClosedChannels cps mq
+  liftLndResult =<< receiveClosedChannels cps cq
+
+receiveActiveChannel ::
+  (LndTest m) =>
+  ChannelPoint ->
+  TChan ((), ChannelEventUpdate) ->
+  m (Either LndError ())
+receiveActiveChannel = this 0
+  where
+    this (attempt :: Integer) cp cq =
+      if attempt > 30
+        then pure $ Left $ LndError "receiveActiveChannel - exceeded"
+        else do
+          x <- readTChanTimeout (MicroSecondsDelay 1000000) cq
+          case channelEvent . snd <$> x of
+            Just (ChannelEventUpdateChannelActiveChannel gcp) ->
+              if cp == gcp
+                then pure $ Right ()
+                else mine1 >> this (attempt + 1) cp cq
+            _ ->
+              mine1 >> this (attempt + 1) cp cq
+
+setupOneChannel :: (LndTest m) => m ()
+setupOneChannel = do
+  alice <- getLndEnv Alice
+  bob <- getLndEnv Bob
+  mq <- dupChannelTChan Bob
+  cq <- dupChannelTChan Alice
+  --
+  -- Open channel from Customer to Merchant
+  --
+  $(logTM) InfoS "SetupOneChannel - opening channel ..."
+  GetInfoResponse merchantPubKey _ _ <-
+    liftLndResult =<< LND.getInfo bob
+  let openChannelRequest =
+        OpenChannel.OpenChannelRequest
+          { OpenChannel.nodePubkey = merchantPubKey,
+            OpenChannel.localFundingAmount = MoneyAmount 20000,
+            OpenChannel.pushSat = Just $ MoneyAmount 1000,
+            OpenChannel.targetConf = Nothing,
+            OpenChannel.satPerByte = Nothing,
+            OpenChannel.private = Nothing,
+            OpenChannel.minHtlcMsat = Nothing,
+            OpenChannel.remoteCsvDelay = Nothing,
+            OpenChannel.minConfs = Nothing,
+            OpenChannel.spendUnconfirmed = Nothing,
+            OpenChannel.closeAddress = Nothing
+          }
+  cp <-
+    liftLndResult
+      =<< LND.openChannelSync alice openChannelRequest
+  liftLndResult =<< receiveActiveChannel cp mq
+  liftLndResult =<< receiveActiveChannel cp cq
+  --
+  -- TODO : this invoice is added and settled to
+  -- raise invoice index to 1 to be able to receive
+  -- notifications about all next invoices
+  -- remove when LND bug will be fixed
+  -- https://github.com/lightningnetwork/lnd/issues/2469
+  --
+  let addInvoiceRequest =
+        AddInvoice.AddInvoiceRequest
+          { AddInvoice.memo = Just "HELLO",
+            AddInvoice.value = MoneyAmount 1000,
+            AddInvoice.expiry = Just $ Seconds 1000
+          }
+  invoice <-
+    liftLndResult =<< LND.addInvoice bob addInvoiceRequest
+  let sendPaymentRequest =
+        SendPayment.SendPaymentRequest
+          { SendPayment.paymentRequest =
+              AddInvoice.paymentRequest invoice,
+            SendPayment.amt =
+              MoneyAmount 1000
+          }
+  void $
+    liftLndResult =<< LND.sendPayment alice sendPaymentRequest
+  $(logTM) InfoS "SetupOneChannel - finished"
+
+receiveInvoice ::
+  ( MonadUnliftIO m,
+    KatipContext m
+  ) =>
+  RHash ->
+  GRPC.Invoice_InvoiceState ->
+  TChan (SubscribeInvoicesRequest, Invoice) ->
+  m (Either LndError ())
+receiveInvoice rh s q = do
+  mx0 <- readTChanTimeout (MicroSecondsDelay 30000000) q
+  let mx = snd <$> mx0
+  $(logTM) InfoS $ logStr $
+    "receiveInvoice - " <> (show mx :: Text)
+  case (\x -> Invoice.rHash x == rh && Invoice.state x == s) <$> mx of
+    Just True -> return $ Right ()
+    Just False -> receiveInvoice rh s q
+    Nothing -> return . Left $ TChanTimeout "receiveInvoice"
 
 spawnLinkChannelWatcher ::
   (KatipContext m, MonadUnliftIO m) =>
